tree-cfg: Move the returns_twice check to be last statement only [PR114301]
When I was checking to making sure that all of the bugs dealing
with the case where gimple_can_duplicate_bb_p would return false was fixed,
I noticed that the code which was checking if a call statement was
returns_twice was checking all call statements rather than just the
last statement. Since calling gimple_call_flags has a small non-zero
overhead due to a few string comparison, removing the uses of it
can have a small performance improvement. In the case of returns_twice
functions calls, will always end the basic-block due to the check in
stmt_can_terminate_bb_p (and others). So checking only the last statement
is a small optimization and will be safe.

Bootstrapped and tested pon x86_64-linux-gnu with no regressions.

	PR tree-optimization/114301
gcc/ChangeLog:

	* tree-cfg.cc (gimple_can_duplicate_bb_p): Check returns_twice
	only on the last call statement rather than all.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>