tree-optimization/79958 - make DSE track multiple paths
DSE currently gives up when the path we analyze forks.  This leads
to multiple missed dead store elimination PRs.  The following fixes
this by recursing for each path and maintaining the visited bitmap
to avoid visiting CFG re-merges multiple times.  The overall cost
is still limited by the same bound, it's just more likely we'll hit
the limit now.  The patch doesn't try to deal with byte tracking
once a path forks but drops info on the floor and only handling
fully dead stores in that case.

	PR tree-optimization/79958
	PR tree-optimization/109087
	PR tree-optimization/100314
	PR tree-optimization/114774
	* tree-ssa-dse.cc (dse_classify_store): New forwarder.
	(dse_classify_store): Add arguments cnt and visited, recurse
	to track multiple paths when we end up with multiple defs.

	* gcc.dg/tree-ssa/ssa-dse-48.c: New testcase.
	* gcc.dg/tree-ssa/ssa-dse-49.c: Likewise.
	* gcc.dg/tree-ssa/ssa-dse-50.c: Likewise.
	* gcc.dg/tree-ssa/ssa-dse-51.c: Likewise.
	* gcc.dg/graphite/pr80906.c: Avoid DSE of last data reference
	in loop.
	* g++.dg/ipa/devirt-24.C: Adjust for extra DSE.
	* g++.dg/warn/Wuninitialized-pr107919-1.C: Use more important
	-O2 optimization level, -O1 regresses.