RISC-V: testsuite: Drop march-string in cmpmemsi/cpymemsi tests
The tests cmpmemsi-1.c and cpymemsi-1.c are execution ("dg-do run")
tests, which does not have any restrictions for the enabled extensions.
Further, no other listed options are required.
Let's drop the options, so that the test can also be executed on
non-f and non-d targets.  However, we need to set options to the
defaults without '-ansi', because the included test file uses the
'asm' keyword, which is not part of ANSI C.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/cmpmemsi-1.c: Drop options.
	* gcc.target/riscv/cpymemsi-1.c: Likewise.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>