Avoid requiring VEC_PERM represenatives
The following plugs one hole where we require a VEC_PERM node
representative unnecessarily.  This is for vect_check_store_rhs
which looks at the RHS and checks whether a constant can be
native encoded.  The fix is to guard that with vect_constant_def
additionally and making vect_is_simple_use forgiving for a missing
SLP_TREE_REPRESENTATIVE when the child is a VEC_PERM node,
initializing the scalar def to error_mark_node.

	* tree-vect-stmts.cc (vect_check_store_rhs): Look at *rhs
	only when it's a vec_constant_def.
	(vect_is_simple_use): When we have no representative for
	an internal node, fill in *op with error_mark_node.