c: Fix for some variably modified types not being recognized [PR114831]
We did not evaluate expressions with variably modified types correctly
in typeof and did not produce warnings when jumping over declarations
using typeof.  After addressof or array-to-pointer decay we construct
new pointer types that have to be marked variably modified if the pointer
target is variably modified.

2024-05-18 Martin Uecker  <uecker@tugraz.at>

	PR c/114831
gcc/c/
	* c-typeck.cc (array_to_pointer_conversion, build_unary_op):
	Propagate flag to pointer target.

gcc/testsuite/
	* gcc.dg/pr114831-1.c: New test.
	* gcc.dg/pr114831-2.c: New test.
	* gcc.dg/gnu23-varmod-1.c: New test.
	* gcc.dg/gnu23-varmod-2.c: New test.