testsuite: Fix typo in torture/vector-{1,2}.c
When making some clean up patches, I happened to find test
cases vector-{1,2}.c are having typo "powerpc64--*-*" in
target selector, which should be powerpc64-*-*.  The reason
why we didn't catch before is that all our testing machines
support VMX insns, so it passes always.  But it would break
if a test machine doesn't support that, so this patch is to
fix it to ensure robustness.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/vector-1.c: Fix typo.
	* gcc.dg/torture/vector-2.c: Likewise.