[testsuite] xfail pr79004 on longdouble64; drop long_double_64bit
Some of the asm opcodes expected by pr79004 depend on
-mlong-double-128 to be output.  E.g., without this flag, the
conditions of patterns @extenddf<mode>2 and extendsf<mode>2 do not
hold, and so GCC resorts to libcalls instead of even trying
rs6000_expand_float128_convert.

Perhaps the conditions are too strict, and they could enable the use
of conversion insns involving __ieee128/_Float128 even with 64-bit
long doubles.

For now, xfail the opcodes that are not available on longdouble64.

While at that, drop long_double_64bit, since it's broken and sort of
redundant.


for  gcc/testsuite/ChangeLog

	PR target/105359
	* gcc.target/powerpc/pr79004.c: Xfail opcodes not available on
	longdouble64.
	* lib/target-supports.exp
	(check_effective_target_long_double_64bit): Drop.
	(add_options_for_long_double_64bit): Likewise.