ada: Fix inlining of fixed-lower-bound array for GNATprove
Inlining in GNATprove may fail on a call to a subprogram with a formal
of an array type with fixed lower bound (a GNAT extension), because the
appropriate conversion is not used. Fix it.

Also fix the function that inserts an unchecked conversion, in cases where
it could skip sliding due to the target type having fixed lower bound.

gcc/ada/

	* inline.adb (Establish_Actual_Mapping_For_Inlined_Call): In the
	case of formal with a fixed lower bounds, insert appropriate
	conversion like in the case of a constrained type.
	* tbuild.adb (Unchecked_Convert_To): Do not skip the conversion
	when it may involve sliding due to a type with fixed lower bound.